pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "streamlit>=1.45.0",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]
